+++ backend/canisters/user/impl/src/model/mod.rs
@@ -5,7 +5,6 @@ pub mod group_chats;
 pub mod local_user_index_event_batch;
 pub mod p2p_swaps;
 pub mod premium_items;
-pub mod referrals;
 pub mod threads_read;
 pub mod token_swaps;
 pub mod user_canister_event_batch;